# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI-Powered Regulatory Intelligence Platform API",
    "description": "Enterprise Compliance Intelligence, Knowledge Graph, Versioning & Audit Engine API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/sources": {
      "get": {
        "tags": [
          "Regulatory Intelligence Domain"
        ],
        "summary": "List Sources",
        "operationId": "list_sources_api_v1_sources_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SourceResponse"
                  },
                  "type": "array",
                  "title": "Response List Sources Api V1 Sources Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Regulatory Intelligence Domain"
        ],
        "summary": "Create Source",
        "operationId": "create_source_api_v1_sources_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sources/{source_id}/trigger": {
      "post": {
        "tags": [
          "Regulatory Intelligence Domain"
        ],
        "summary": "Trigger Fetch",
        "operationId": "trigger_fetch_api_v1_sources__source_id__trigger_post",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Trigger Fetch Api V1 Sources  Source Id  Trigger Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/tasks": {
      "get": {
        "tags": [
          "Compliance Workflow Domain"
        ],
        "summary": "List Tasks",
        "operationId": "list_tasks_api_v1_tasks_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "assignee",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Assignee"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskResponse"
                  },
                  "title": "Response List Tasks Api V1 Tasks Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Compliance Workflow Domain"
        ],
        "summary": "Create Task",
        "operationId": "create_task_api_v1_tasks_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}": {
      "put": {
        "tags": [
          "Compliance Workflow Domain"
        ],
        "summary": "Update Task",
        "operationId": "update_task_api_v1_tasks__task_id__put",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SourceCreate": {
        "properties": {
          "authority_name": {
            "type": "string",
            "title": "Authority Name"
          },
          "feed_url": {
            "type": "string",
            "title": "Feed Url"
          },
          "feed_type": {
            "type": "string",
            "title": "Feed Type",
            "default": "RSS"
          },
          "fetch_schedule": {
            "type": "string",
            "title": "Fetch Schedule",
            "default": "HOURLY"
          },
          "region": {
            "type": "string",
            "title": "Region"
          },
          "sector": {
            "type": "string",
            "title": "Sector"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "ACTIVE"
          }
        },
        "type": "object",
        "required": [
          "authority_name",
          "feed_url",
          "region",
          "sector"
        ],
        "title": "SourceCreate"
      },
      "SourceResponse": {
        "properties": {
          "authority_name": {
            "type": "string",
            "title": "Authority Name"
          },
          "feed_url": {
            "type": "string",
            "title": "Feed Url"
          },
          "feed_type": {
            "type": "string",
            "title": "Feed Type",
            "default": "RSS"
          },
          "fetch_schedule": {
            "type": "string",
            "title": "Fetch Schedule",
            "default": "HOURLY"
          },
          "region": {
            "type": "string",
            "title": "Region"
          },
          "sector": {
            "type": "string",
            "title": "Sector"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "ACTIVE"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "last_fetched_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Fetched At"
          }
        },
        "type": "object",
        "required": [
          "authority_name",
          "feed_url",
          "region",
          "sector",
          "id",
          "last_fetched_at"
        ],
        "title": "SourceResponse"
      },
      "TaskCreate": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "assignee": {
            "type": "string",
            "title": "Assignee"
          },
          "reviewer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer"
          },
          "priority": {
            "type": "string",
            "title": "Priority",
            "default": "HIGH"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "NEEDS_REVIEW"
          },
          "due_date": {
            "type": "string",
            "format": "date",
            "title": "Due Date"
          },
          "regulation_id": {
            "type": "string",
            "title": "Regulation Id"
          },
          "control_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Control Code"
          }
        },
        "type": "object",
        "required": [
          "title",
          "assignee",
          "due_date",
          "regulation_id"
        ],
        "title": "TaskCreate"
      },
      "TaskResponse": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "assignee": {
            "type": "string",
            "title": "Assignee"
          },
          "reviewer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer"
          },
          "priority": {
            "type": "string",
            "title": "Priority",
            "default": "HIGH"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "NEEDS_REVIEW"
          },
          "due_date": {
            "type": "string",
            "format": "date",
            "title": "Due Date"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "regulation_id": {
            "type": "string",
            "title": "Regulation Id"
          },
          "control_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Control Code"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "regulation_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Regulation Title"
          }
        },
        "type": "object",
        "required": [
          "title",
          "assignee",
          "due_date",
          "id",
          "regulation_id",
          "created_at",
          "updated_at"
        ],
        "title": "TaskResponse"
      },
      "TaskUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "assignee": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee"
          },
          "reviewer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer"
          },
          "priority": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "due_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          }
        },
        "type": "object",
        "title": "TaskUpdate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}